+++ src/__tests__/pages/pupilHome.test.tsx
@@ -1,0 +1,16 @@
+import { screen, waitFor } from "@testing-library/react";
+
+import PupilHome from "../../pages/beta/pupil";
+import renderWithProviders from "../__helpers__/renderWithProviders";
+
+describe("pages/pupilHome.tsx", () => {
+  it("Renders correct title ", async () => {
+    renderWithProviders(<PupilHome />);
+
+    await waitFor(() => {
+      expect(screen.getByRole("heading", { level: 1 }).textContent).toBe(
+        "Pupil Home"
+      );
+    });
+  });
+});