+++ src/components/Card/Card.tsx
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 
+import border, { BorderProps } from "../../styles/utils/border";
 import Flex from "../Flex/Flex";
 
 /**
@@ -7,15 +8,16 @@ import Flex from "../Flex/Flex";
  * and 'background' util types
  */
 
-const Card = styled(Flex)`
-  min-height: 200px;
-  border-radius: 12px;
+const Card = styled(Flex)<BorderProps>`
+  position: relative;
   flex-grow: 1;
+  ${border}
 `;
 
 Card.defaultProps = {
   pa: 24,
   flexDirection: "column",
+  borderRadius: 12,
 };
 
 export default Card;