+++ src/components/Flex/Flex.tsx
@@ -5,8 +5,9 @@ import background from "../../styles/utils/background";
 import position from "../../styles/utils/position";
 import spacing from "../../styles/utils/spacing";
 import flex, { FlexCssProps } from "../../styles/utils/flex";
+import border, { BorderProps } from "../../styles/utils/border";
 
-export type FlexProps = FlexCssProps & BoxProps;
+export type FlexProps = FlexCssProps & BoxProps & BorderProps;
 /**
  * Flex sets `display: flex;` and exposes various flex props, along with Box
  * props
@@ -16,6 +17,7 @@ const Flex = styled(Box)<FlexProps>`
   ${spacing}
   ${background}
   ${position}
+  ${border}
 `;
 
 export default Flex;