+++ src/components/SiteHeader/SiteHeader.tsx
@@ -2,19 +2,11 @@ import { FC } from "react";
 import Link from "next/link";
 import styled from "styled-components";
 
-import background from "../../styles/utils/background";
 import Flex, { FlexProps } from "../Flex";
-import Icon from "../Icon";
 import flex from "../../styles/utils/flex";
 import P, { Span } from "../Typography";
-
-const StyledSiteHeader = styled.header<FlexProps>`
-  width: 100%;
-  min-height: 72px;
-  padding: 12px;
-  ${flex}
-  ${background}
-`;
+import Icon from "../Icon";
+import FixedHeader from "../FixedHeader";
 
 const HomeLink = styled.a<FlexProps>`
   ${flex}
@@ -22,12 +14,7 @@ const HomeLink = styled.a<FlexProps>`
 
 const SiteHeader: FC = () => {
   return (
-    <StyledSiteHeader
-      background={"madangGreen"}
-      justifyContent={["left", "space-between"]}
-      alignItems={["flex-start", "center"]}
-      flexDirection={["column", "row"]}
-    >
+    <FixedHeader background={"madangGreen"}>
       <Link href={"/"} passHref>
         <HomeLink alignItems="center">
           <Icon name="Home" size={32} mr={8} />
@@ -41,17 +28,20 @@ const SiteHeader: FC = () => {
           </Span>
         </HomeLink>
       </Link>
-      <Flex alignItems={"center"} mt={[16, 0]}>
+      <Flex alignItems={"center"} mt={[16, 0, 0]}>
         <P>
+          <Link href={"/beta/onboarding"}>Join the Beta</Link>
+        </P>
+        <P ml={[16, 48, 48]}>
           <Link href={"https://classroom.thenational.academy/"}>Classroom</Link>
         </P>
-        <P ml={[48]}>
+        <P ml={[16, 48, 48]}>
           <Link href={"https://teachers.thenational.academy/"}>
             Teacher Hub
           </Link>
         </P>
       </Flex>
-    </StyledSiteHeader>
+    </FixedHeader>
   );
 };