+++ src/styles/theme/types.ts
@@ -3,7 +3,8 @@ import { z } from "zod";
 
 import { BadgeConfig } from "../../components/Badge";
 import { CheckboxConfig } from "../../components/Checkbox";
-import { SelectListBoxConfig } from "../../components/DropDownSelect/ListBox";
+import { SelectListBoxConfig } from "../../components/DropdownSelect/ListBox";
+import { HeaderConfig } from "../../components/FixedHeader/FixedHeader";
 import { LessonControlConfig } from "../../components/LessonControl";
 import { ToggleStyleConfig } from "../../components/Toggle/Toggle";
 
@@ -90,6 +91,7 @@ type InputConfig = {
   height: string;
   borderRadius: string;
   borderWidth: string;
+  fontSize: string;
   states: {
     default: InputStateConfig;
     active: InputStateConfig;
@@ -108,6 +110,7 @@ type InputStateConfig = {
 };
 
 export type OakTheme = {
+  header: HeaderConfig;
   name: string;
   colors: Record<OakColorName, ColorValue>;
   contrastColors: Record<OakColorName, OakColorName>;